Adding save method to setTranslation function (#34)
I've been using the functions `forgetTranslation` and `setTranslations` which work perfectly without using the `save()` method on a model, but for `setTranslation` you need to save after setting the translation. 

I've been trying to figure out what the reason could be why it wouldn't save, but I found out after a few minutes of searching you've to add the `save()` method after you called `setTranslation`, which wasn't clear to me at the first place.

So in my opinion, since I know its on the top of the readme, it might be better to also add it close to the function?