Set a default translation for the translatable attr.
This PR is aimed to provide setter functionality for default translatable fields. Without this method, when you want to save an Eloquent model, "HasTranslations" trait is appliying '"' on the fields and not converting json. So this setAttribute method will automatically convert to json column when saving the model without required to cal setTranslation method