Update HasTranslations.php
json_decode not always returns an array and getTranslaction signature need an array to be returned.

I suggest to return an empty array if json_decode has failed of throw some custom Exception that the application can catch and decide what to do.

I am working with data saved in old legacy database that, in some records, the json data is not valid. It is not a big matter, so we could just render some "info not available right now". Now, the app is crashing.

I can create my own Trait expanding this one, but maybe you want to value another solution.

Cheers